+++ code/core/src/stringview.h
@@ -1,0 +1,25 @@
+/*
+ * variant.h
+ *
+ *  Created on: Jan 31, 2018
+ *      Author: Dusteh
+ */
+
+#pragma once
+
+#include <string>
+#include <typeinfo>
+#include <memory>
+#include <type_traits>
+
+namespace bricks
+{
+    namespace core
+    {
+        class StringView
+        {
+            
+        };
+
+    }
+}
