+++ code/text/CMakeLists.txt
@@ -3,15 +3,11 @@ cmake_minimum_required(VERSION 3.0)
 # project configuration
 project(textbrick VERSION 0.1 LANGUAGES CXX)
 add_library(textbrick
-    text.cpp
-    textutils.cpp
+    src/text.cpp
+    src/textutils.cpp
 )
 
-target_include_directories(textbrick INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
-
-set_target_properties(textbrick PROPERTIES PUBLIC_HEADER
-    "text.h;textutils.h"
-)
+target_include_directories(textbrick INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/src)
 
 install(TARGETS  textbrick
         RUNTIME  DESTINATION ${INSTALL_SUBMODULES_RUN}
@@ -20,3 +16,6 @@ install(TARGETS  textbrick
         INCLUDES DESTINATION ${INSTALL_SUBMODULES_INC}
         PUBLIC_HEADER DESTINATION ${INSTALL_SUBMODULES_INC}/text/
 )
+
+install(DIRECTORY src/ DESTINATION ${INSTALL_SUBMODULES_INC}/core/
+    FILES_MATCHING PATTERN "*.h")
