--- src/main.cpp
@@ -1,18 +1,0 @@
-/*
- * main.cpp
- *
- *  Created on: Jan 27, 2018
- *      Author: Dusteh
- */
-
-#include "core/variant.h"
-#include "text/textutils.h"
-#include "text/text.h"
-
-#include <iostream>
-#include <string>
-int main()
-{
-	// dummy main for dev purposes
-	return 0;
-}
