--- unit_tests/codeblocks/appbricks_tests.cbp
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6" />
-	<Project>
-		<Option title="appbricks_tests" />
-		<Option pch_mode="2" />
-		<Option compiler="gcc" />
-		<Build>
-			<Target title="Debug">
-				<Option output="bin/Debug/appbricks_tests" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/Debug/" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-g" />
-					<Add option="-std=gnu++14" />
-					<Add option="-DGTEST_OS_WINDOWS_MINGW" />
-				</Compiler>
-				<Linker>
-					<Add library="../libs/libgtest.a" />
-				</Linker>
-			</Target>
-			<Target title="Release">
-				<Option output="bin/Release/appbricks_tests" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/Release/" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-O2" />
-				</Compiler>
-				<Linker>
-					<Add option="-s" />
-				</Linker>
-			</Target>
-		</Build>
-		<Compiler>
-			<Add option="-Wall" />
-			<Add option="-fexceptions" />
-		</Compiler>
-		<Unit filename="../main.cpp" />
-		<Unit filename="../test_text.cpp" />
-		<Unit filename="../test_variant.cpp" />
-		<Unit filename="../../src/core/variant.h" />
-		<Unit filename="../../src/text/text.h" />
-		<Unit filename="../../src/text/textutils.cpp" />
-		<Unit filename="../../src/text/textutils.h" />
-		<Extensions>
-			<code_completion />
-			<envvars />
-			<debugger />
-		</Extensions>
-	</Project>
-</CodeBlocks_project_file>
