--- unit_tests/visualstudio/appbricks_tests.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "appbricks_tests", "appbricks_tests.vcxproj", "{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Debug|x64.ActiveCfg = Debug|x64
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Debug|x64.Build.0 = Debug|x64
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Debug|x86.ActiveCfg = Debug|Win32
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Debug|x86.Build.0 = Debug|Win32
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Release|x64.ActiveCfg = Release|x64
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Release|x64.Build.0 = Release|x64
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Release|x86.ActiveCfg = Release|Win32
-		{B87363D9-7AC3-449B-A1B0-A59E0A2C7837}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
